requirements.txt
Pillow>=10.0
rembg[cpu]>=2.0.55
onnxruntime>=1.16
